+++ src/eval.rs
@@ -3,9 +3,8 @@ use std::ffi::CString;
 use savvy_ffi::{R_ParseEvalString, R_compute_identical, Rboolean_TRUE, SEXP};
 
 use crate::{
-    Sexp,
     protect::{self},
-    unwind_protect,
+    unwind_protect, Sexp,
 };
 
 /// A result of a function call. Since the result does not yet belong to any